requirements_conda.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
#_libgcc_mutex=0.1=main
aenum=2.2.1=py37_0
alive-progress=1.6.1=pypi_0
anagate=0.1.0=pypi_0
argparse=1.2.1=pypi_0
astroml=0.4.1=pypi_0
astropy=4.0.1=pypi_0
attrs=19.3.0=py_0
backcall=0.1.0=py37_0
basil-daq=3.0.1=pypi_0
bitarray=1.2.1=py37h7b6447c_0
blas=1.0=mkl
bleach=3.1.4=py_0
blosc=1.16.3=hd408876_0
boto=2.39.0=pypi_0
branca=0.4.1=py_0
bzip2=1.0.8=h7b6447c_0
ca-certificates=2020.4.5.2=hecda079_0
canlib=1.15.483=pypi_0
celluloid=0.2.0=pypi_0
certifi=2020.4.5.2=py37hc8dfbb8_0
cffi=1.14.0=py37he30daa8_1
chardet=3.0.4=py37_1003
colorama=0.4.3=pypi_0
coloredlogs=10.0=pypi_0
conda=4.8.3=py37hc8dfbb8_1
conda-package-handling=1.6.1=py37h7b6447c_0
contextlib2=0.6.0.post1=py_0
cryptography=2.9.2=py37h1ba5d50_0
cycler=0.10.0=py37_0
cython=0.29.17=py37he6710b0_0
dbus=1.13.14=hb2f20db_0
dcscontroller=1.1.0=dev_0
decorator=4.4.2=py_0
defusedxml=0.6.0=py_0
dill=0.3.1.1=pypi_0
edx-opaque-keys=2.1.0=pypi_0
entrypoints=0.3=py37_0
enum34=1.1.6=pypi_0
expat=2.2.6=he6710b0_0
folium=0.11.0=py_0
fontconfig=2.13.0=h9420a91_0
freetype=2.9.1=h8a8886c_1
glib=2.63.1=h3eb4bd4_1
gmp=6.1.2=h6c8ec71_1
gnupg=2.3.1=pypi_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb31296c_0
hdf5=1.10.4=hb1b8bf9_0
http-parser=0.9.0=pypi_0
humanfriendly=4.18=pypi_0
icu=58.2=he6710b0_3
idna=2.8=pypi_0
importlib-metadata=1.6.0=py37_0
importlib_metadata=1.6.0=0
intel-openmp=2019.4=243
ipykernel=5.1.4=py37h39e3cac_0
ipython=7.13.0=py37h5ca1d4c_0
ipython_genutils=0.2.0=py37_0
ipywidgets=7.5.1=py_0
jedi=0.17.0=py37_0
jinja2=2.11.2=py_0
joblib=0.14.1=pypi_0
jpeg=9b=h024ee3a_2
jsonschema=3.2.0=py37_0
jupyter=1.0.0=py37_7
jupyter_client=6.1.3=py_0
jupyter_console=6.1.0=py_0
jupyter_core=4.6.3=py37_0
kafe=1.3.1=pypi_0
kiwisolver=1.2.0=py37hfd86e86_0
latexmk-py=0.4=pypi_0
libcst=0.3.10=pypi_0
libedit=3.1.20181209=hc058e9b_0
libffi=3.3=he6710b0_1
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libsodium=1.0.16=h1bed415_0
libstdcxx-ng=9.1.0=hdf63c60_0
libuuid=1.0.3=h1bed415_2
libxcb=1.13=h1bed415_1
libxml2=2.9.9=hea5a465_1
llvmlite=0.30.0=py37hd408876_0
lxml=4.4.2=pypi_0
lz4-c=1.8.1.2=h14c3975_0
lzo=2.10=h7b6447c_2
markupsafe=1.1.1=py37h7b6447c_0
matplotlib=3.1.3=py37_0
matplotlib-base=3.1.3=py37hef1b27d_0
mistune=0.8.4=py37h7b6447c_0
mkl=2019.4=243
mkl-service=2.3.0=py37he904b0f_0
mkl_fft=1.0.15=py37ha843d7b_0
mkl_random=1.1.0=py37hd6b4f25_0
mock=4.0.2=py_0
multiprocess=0.70.9=pypi_0
mypy-extensions=0.4.3=pypi_0
nbconvert=5.6.1=py37_0
nbformat=5.0.6=py_0
ncurses=6.2=he6710b0_1
nose=1.2.1=pypi_0
notebook=6.0.3=py37_0
numba=0.46.0=py37h962f231_0
numexpr=2.7.1=py37h423224d_0
numpy=1.17.4=py37hc1035e2_0
numpy-base=1.17.4=py37hde5b4d6_0
opcua=0.98.9=pypi_0
openssl=1.1.1g=h516909a_0
pandas=0.25.3=pypi_0
pandoc=2.2.3.2=0
pandocfilters=1.4.2=py37_1
parso=0.7.0=py_0
path-py=2.4.1=pypi_0
pathlib=1.0.1=pypi_0
pbr=5.4.5=pypi_0
pcre=8.43=he6710b0_0
pexpect=4.8.0=py37_0
pickleshare=0.7.5=py37_0
pillow=7.1.2=pypi_0
pip=20.2.4=pypi_0
progressbar-latest=2.4=pypi_0
prometheus_client=0.7.1=py_0
prompt-toolkit=3.0.5=py_0
prompt_toolkit=3.0.5=0
psutil=5.7.0=pypi_0
ptyprocess=0.6.0=py37_0
pycosat=0.6.3=py37h7b6447c_0
pycparser=2.20=py_0
pygit2=0.26.4=pypi_0
pygments=2.6.1=py_0
pymongo=3.10.1=pypi_0
pyopenssl=19.1.0=py37_0
pyparsing=2.4.7=py_0
pypdf2=1.26.0=pypi_0
pyqt=5.9.2=py37h05f1152_2
pyqtgraph=0.10.0=pypi_0
pyrsistent=0.16.0=py37h7b6447c_0
pyserial=3.4=py37_0
pysocks=1.7.1=py37_0
pytables=3.6.1=py37h71ec239_0
python=3.7.5=h0371630_0
python-bidi=0.4.2=pypi_0
python-can=3.3.2=pypi_0
python-dateutil=2.8.1=py_0
python_abi=3.7=1_cp37m
pytz=2020.1=py_0
pyvisa=1.10.1=pypi_0
pyvisa-py=0.3.1=pypi_0
pyyaml=5.3.1=py37h7b6447c_0
pyzmq=18.1.1=py37he6710b0_0
qt=5.9.7=h5867ecd_1
qtawesome=0.6.1=py_0
qtconsole=4.7.4=py_0
qtpy=1.9.0=py_0
readline=7.0=h7b6447c_5
reportlab=3.5.42=pypi_0
requests=2.22.0=pypi_0
ruamel_yaml=0.15.87=py37h7b6447c_0
scikit-learn=0.22.2.post1=pypi_0
scipy=1.4.1=py37h0b6359f_0
seaborn=0.9.0=pypi_0
send2trash=1.5.0=py37_0
setuptools=49.1.3=pypi_0
sip=4.19.8=py37hf484d3e_0
six=1.15.0=pypi_0
snappy=1.1.7=hbae5bb6_3
sqlite=3.31.1=h62c20be_1
stevedore=1.32.0=pypi_0
tbb=2020.0=hfd86e86_0
termcolor=1.1.0=pypi_0
terminado=0.8.3=py37_0
testpath=0.4.4=py_0
timeout-decorator=0.4.1=pypi_0
tk=8.6.8=hbc83047_0
toml=0.10.1=pypi_0
tornado=6.0.4=py37h7b6447c_1
tqdm=4.46.0=py_0
traitlets=4.3.3=py37_0
typing-extensions=3.7.4.3=pypi_0
typing-inspect=0.6.0=pypi_0
unimport=0.3.0=pypi_0
uptime=3.0.1=pypi_0
urllib3=1.25.9=pypi_0
verboselogs=1.7=py37_0
vulture=2.1=pypi_0
wcwidth=0.1.9=py_0
webencodings=0.5.1=py37_1
wheel=0.34.2=py37_0
widgetsnbextension=3.5.1=py37_0
wrapt=1.12.1=pypi_0
wrapt-timeout-decorator=1.3.1=pypi_0
xz=5.2.5=h7b6447c_0
yaml=0.1.7=had09818_2
zeromq=4.3.1=he6710b0_3
zipp=3.1.0=py_0
zlib=1.2.11=h7b6447c_3
zstd=1.3.7=h0b5b093_0
